refactor: restructure `primitives` crate
