+++ chain-signatures/primitives/Cargo.toml
@@ -4,6 +4,9 @@ version.workspace = true
 edition = "2021"
 
 [dependencies]
+alloy.workspace = true
+# TODO: probably should transition to thiserror
+anyhow.workspace = true
 ciborium.workspace = true
 hex.workspace = true
 k256.workspace = true
